requirements.txt
httpx==0.28.1
openai==1.102.0
python-dotenv==1.1.1